+++ CHANGELOG.md
@@ -29,6 +29,7 @@ adheres to [Semantic Versioning](http://semver.org/).
 - Migrated from Travis-CI to GitHub Actions.
 - Simplified configuration for backbone models and losses.
 - Simplified contributing documentation.
+- Uniform kernel size for LNCC loss.
 
 ### Fixed